+++ cmd/sessionhub/main.go
@@ -14,7 +14,7 @@ import (
 )
 
 var (
-	version   = "0.1.22"
+	version   = "0.2.0"
 	commit    = "none"
 	buildDate = "unknown"
 )